+++ src/modules/auth/auth.interface.ts
@@ -1,3 +1,13 @@
-export interface UserDataPayload {
+export interface UserAccessTokenPayload {
+    login: string;
+    capabilities?: {
+        canAdd?: boolean;
+        canEdit?: boolean;
+        canDelete?: boolean;
+    }
+    isAdmin?: boolean;
+}
+
+export interface UserRefreshTokenPayload {
     login: string;
 }