+++ src/modules/auth/auth.module.ts
@@ -5,11 +5,13 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { USER_ACTION_MODEL, USER_MODEL } from '../../constants/models.constant';
 import { UserRepository } from '../../mongodb/repositories/user.repository';
 import { UserActionRepository } from '../../mongodb/repositories/user.action.repository';
+import { RedisModule } from '../redis/redis.module';
 
 @Module({
     imports: [
         MongooseModule.forFeature([USER_MODEL, USER_ACTION_MODEL]),
-        JwtManagerModule
+        JwtManagerModule,
+        RedisModule
     ],
     providers: [AuthService, UserRepository, UserActionRepository],
     exports: [AuthService]