+++ src/modules/auth/auth.service.spec.ts
@@ -10,6 +10,7 @@ import { REDIS_CLIENT } from '../redis/redis.constants';
 import { MailManagerService } from '../mail-manager/mail-manager.service';
 import { UserRepository } from '../../mongodb/repositories/user.repository';
 import { UserActionRepository } from '../../mongodb/repositories/user.action.repository';
+import { RedisService } from '../redis/redis.service';
 
 // https://betterprogramming.pub/testing-controllers-in-nestjs-and-mongo-with-jest-63e1b208503c
 // https://stackoverflow.com/questions/74110962/please-make-sure-that-the-argument-databaseconnection-at-index-0-is-available
@@ -59,7 +60,7 @@ describe('AuthService', () => {
                     useValue: {}
                 },
                 {
-                    provide: REDIS_CLIENT,
+                    provide: RedisService,
                     useValue: {}
                 },
                 {
@@ -110,7 +111,7 @@ describe('AuthService', () => {
             const jwtCookie = 'some cookie';
 
             // When
-            jest.spyOn(jwtManagerService, 'decodeUserData').mockReturnValueOnce('some user name');
+            jest.spyOn(jwtManagerService, 'decodeUserData').mockReturnValueOnce({ login: 'some user login' });
             jest.spyOn(userRepository, 'findByLogin').mockReturnValueOnce(null);
 
             // Then