+++ src/modules/meal/meal.controller.ts
@@ -1,6 +1,10 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpCode, Param, Post, Put, UseGuards } from '@nestjs/common';
 import { MealService } from './meal.service';
 import { CreateMealBodyDto, EditMealBodyDto } from './meal.dto';
+import { AuthenticationGuard } from '../../guards/authentication.guard';
+import { CreationGuard } from '../../guards/creation.guard';
+import { EditionGuard } from '../../guards/edition.guard';
+import { DeletionGuard } from '../../guards/deletion.guard';
 
 @Controller('meals')
 export class MealController {
@@ -20,6 +24,7 @@ export class MealController {
 
     @Post('/create')
     @HttpCode(201)
+    @UseGuards(AuthenticationGuard)
     public async createMeal(@Body() body: CreateMealBodyDto) {
         const { data } = body;
 
@@ -28,12 +33,14 @@ export class MealController {
 
     @Delete('/:id')
     @HttpCode(204)
+    @UseGuards(AuthenticationGuard)
     public async deleteMeal(@Param('id') id: string) {
         return await this.mealService.delete(id);
     }
 
     @Put('/:id')
     @HttpCode(200)
+    @UseGuards(AuthenticationGuard)
     public async updateMeal(@Param('id') id: string, @Body() body: EditMealBodyDto) {
         const { data } = body;
 
@@ -42,6 +49,7 @@ export class MealController {
 
     @Post('/:id/create')
     @HttpCode(200)
+    @UseGuards(AuthenticationGuard, CreationGuard)
     public async confirmCreatingMeal(@Param('id') id: string, @Body() body) {
         const { authenticatedUser } = body;
 
@@ -50,6 +58,7 @@ export class MealController {
 
     @Post('/:id/edit')
     @HttpCode(200)
+    @UseGuards(AuthenticationGuard, EditionGuard)
     public async confirmEditingMeal(@Param('id') id: string, @Body() body) {
         const { authenticatedUser } = body;
 
@@ -58,6 +67,7 @@ export class MealController {
 
     @Post('/:id/delete')
     @HttpCode(200)
+    @UseGuards(AuthenticationGuard, DeletionGuard)
     public async confirmDeletingMeal(@Param('id') id: string, @Body() body) {
         const { authenticatedUser } = body;