+++ src/modules/user/user.controller.ts
@@ -1,8 +1,10 @@
-import { Body, Controller, HttpCode, Post, Response, Param } from '@nestjs/common';
+import { Body, Controller, HttpCode, Post, Response, Param, UseGuards, Request } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto, UserDto, UserLoginDto } from './user.dto';
 import { CapabilityType } from './user.types';
 import { UserRepository } from '../../mongodb/repositories/user.repository';
+import { AuthenticationGuard } from '../../guards/authentication.guard';
+import { CapabilityGuard } from '../../guards/capability.guard';
 
 @Controller('users')
 export class UserController {
@@ -11,15 +13,18 @@ export class UserController {
                 private readonly userRepository: UserRepository) {}
 
     @Post('/login')
-    @HttpCode(200)
+    @HttpCode(204)
     public async login(@Body() loginBody: UserLoginDto, @Response({ passthrough: true }) res) {
         return await this.userService.loginUser(loginBody, res);
     }
 
     @Post('/logout')
-    @HttpCode(200)
-    public async logout(@Response({ passthrough: true }) res) {
-        return await this.userService.logoutUser(res);
+    @HttpCode(205)
+    public async logout(@Request() req, @Response({ passthrough: true }) res) {
+        const { accessToken, refreshToken } = req.cookies;
+        const { login } = req.body;
+
+        return await this.userService.logoutUser(res, login, accessToken, refreshToken);
     }
 
     @Post('/create')
@@ -30,6 +35,7 @@ export class UserController {
 
     @Post('/:login/grant/:capability')
     @HttpCode(200)
+    @UseGuards(AuthenticationGuard, CapabilityGuard)
     public async grantPermission(@Body() body, @Param('login') login: string, @Param('capability') capability: CapabilityType) {
         const forUser = await this.userRepository.findByLogin(login) as unknown as UserDto;
         const { authenticatedUser } = body;
@@ -39,6 +45,7 @@ export class UserController {
 
     @Post('/:login/deny/:capability')
     @HttpCode(200)
+    @UseGuards(AuthenticationGuard, CapabilityGuard)
     public async denyPermission(@Body() body, @Param('login') login: string, @Param('capability') capability: CapabilityType) {
         const forUser = await this.userRepository.findByLogin(login) as unknown as UserDto;
         const { authenticatedUser } = body;