+++ src/modules/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { CreateUserDto, UserDto, UserLoginDto } from './user.dto';
 import * as bcrypt from 'bcrypt';
 import { UserDocument } from '../../mongodb/documents/user.document';
@@ -7,13 +7,14 @@ import { JwtManagerService } from '../jwt-manager/jwt-manager.service';
 import { BadRequestException } from '../../exceptions/bad-request.exception';
 import { LoggerService } from '../logger/logger.service';
 import { NotFoundException } from '../../exceptions/not-found.exception';
-import { Redis } from 'ioredis';
 import { CapabilityType } from './user.types';
 import { MailManagerService } from '../mail-manager/mail-manager.service';
 import { UserActionDocument } from '../../mongodb/documents/user-action.document';
 import { ForbiddenException } from '../../exceptions/forbidden-exception';
 import { UserRepository } from '../../mongodb/repositories/user.repository';
 import { UserActionRepository } from '../../mongodb/repositories/user.action.repository';
+import { RedisService } from '../redis/redis.service';
+import { Response } from 'express';
 
 @Injectable()
 export class UserService {
@@ -21,13 +22,13 @@ export class UserService {
     constructor(
         private readonly userRepository: UserRepository,
         private readonly userActionRepository: UserActionRepository,
-        @Inject('REDIS_CLIENT') private readonly redis: Redis,
+        private readonly redisService: RedisService,
         private readonly jwtManagerService: JwtManagerService,
         private readonly loggerService: LoggerService,
         private readonly mailManagerService: MailManagerService
     ) {}
 
-    async loginUser(userLoginDto: UserLoginDto, res): Promise<UserDocument> {
+    async loginUser(userLoginDto: UserLoginDto, res: Response) {
         const { login, password } = userLoginDto;
         const context = 'UserService/loginUser';
         const user = await this.userRepository.findByLogin(login);
@@ -53,18 +54,27 @@ export class UserService {
             throw new BadRequestException(context, message);
         }
 
-        const jwt = await this.jwtManagerService.encodeUserData({ login });
-        res.cookie('jwt', jwt, { httpOnly: true });
+        // @ts-ignore
+        const accessToken = await this.jwtManagerService.generateAccessToken({ login, isAdmin: user.isAdmin, capabilities: user.capabilities });
+        const refreshToken = await this.jwtManagerService.generateRefreshToken({ login });
+
+        await this.redisService.setTokens(login, accessToken, refreshToken);
+        res.cookie('accessToken', accessToken, { httpOnly: true });
+        res.cookie('refreshToken', refreshToken, { httpOnly: true });
+
         const message = `User "${login}" has been successfully logged in`;
         this.loggerService.info(context, message);
-
-        return user;
     }
 
-    async logoutUser(res): Promise<null> {
-        res.clearCookie('jwt');
+    async logoutUser(res: Response, login: string, accessToken: string, refreshToken: string): Promise<void> {
+        try {
+            await this.redisService.unsetTokens(login, accessToken, refreshToken);
+        } catch (err: any) {
+            throw err;
+        }
 
-        return null;
+        res.clearCookie('accessToken');
+        res.clearCookie('refreshToken');
     }
 
     async createUser(createUserDto: CreateUserDto): Promise<UserDocument> {